Revert "Allow payment methods to use tailored workflows"
