--- frontend/src/actions/payment/default.ts
@@ -1,10 +1,0 @@
-import type { PaymentActionContext, PaymentMethodAction } from './types'
-
-export const createDefaultAction = (context: PaymentActionContext): PaymentMethodAction => ({
-  handleSelection: ({ method, currency }) => {
-    context.openMethodUrl(method, currency)
-  },
-  handleCurrencySelection: ({ method, currency }) => {
-    context.openMethodUrl(method, currency)
-  },
-})
