--- frontend/src/actions/payment/toss.ts
@@ -1,75 +1,0 @@
-import { resolveDeepLink } from '@/utils/deepLink'
-
-import type { PaymentActionContext, PaymentMethodAction } from './types'
-
-const ensureTossDeepLink = async (context: PaymentActionContext): Promise<string | null> => {
-  const loaded = await context.ensurePaymentInfoLoaded()
-
-  if (!loaded) {
-    return null
-  }
-
-  const info = context.getDeepLinkInfo('toss')
-
-  if (!info) {
-    console.error('Missing Toss payment info payload')
-    return null
-  }
-
-  try {
-    return resolveDeepLink('toss', info)
-  } catch (error) {
-    console.error(error)
-    return null
-  }
-}
-
-const openTossFallback = (context: PaymentActionContext, url: string) => {
-  context.showDeepLinkPopup('not-mobile', 'toss')
-  context.openUrlInNewTab(url)
-}
-
-const launchTossDeepLink = async (
-  context: PaymentActionContext,
-  url: string,
-) => {
-  context.setDeepLinkChecking(true)
-
-  try {
-    const launchMonitor = context.waitForDeepLinkResult(2000)
-    const didNavigate = context.navigateToDeepLink(url)
-
-    if (!didNavigate) {
-      await launchMonitor
-      return
-    }
-
-    const didLaunch = await launchMonitor
-
-    if (!didLaunch) {
-      context.showDeepLinkPopup('not-installed', 'toss')
-    }
-  } finally {
-    context.setDeepLinkChecking(false)
-  }
-}
-
-const runTossWorkflow = async (context: PaymentActionContext) => {
-  const deepLink = await ensureTossDeepLink(context)
-
-  if (!deepLink) {
-    return
-  }
-
-  if (!context.isMobileDevice()) {
-    openTossFallback(context, deepLink)
-    return
-  }
-
-  await launchTossDeepLink(context, deepLink)
-}
-
-export const createTossAction = (context: PaymentActionContext): PaymentMethodAction => ({
-  handleSelection: () => runTossWorkflow(context),
-  handleCurrencySelection: () => runTossWorkflow(context),
-})
